Fix rax_auth token caching
The atom handlers can use the keystone-ish rax_auth authentication with
an atompub server that supports it (like AtomHopper).

The auth code is supposed to fetch an auth token and use it until it
expires, then re-auth. This patch fixes a bug where yagi re-auths
every batch because the token caching mechanism is broken.