Improve config error handling and timeouts.
Catch NoSectionError in yagi.config.get if there is a missing config.
There are several optional configs, and these errors keep getting missed,
causing production issues. Now it will return None, even if the value is
in a section which is not there 9the same as it does for a missing value in
an existing section.

Also, add missing timeout for handlers which make http calls (AtomPub, CufPub)
It would be nice to drom httplib2  for these and use Requests, but that is more
involved, and for another day.