Fix deterministic ID generation in CUF handler.
Message ids for CUF format messages are supposed to be
deterministicly generated via the UUID5 algorithm, to prevent
duplicate copies upstream if the same verified message is
sent to Yagi more than once. The id was being properly generated,
but was being ignored and replaced with a random UUID.

This fixes that issue.